Grab and Clean Changes From Prototype Wizard Extension Hookup (#26)
Summary of important changes:
- Rename wizard extension types to have `Config` on the end, and change
`WizardBasedExtension` to be a class which makes use of the `Config`
types. Am diong so in order to prepare for information being sent
declaratively and directly by the service extensions.
- Propertly activate wizard extensions.
- Create a AgentUserInput.d.ts file which contains all the
`IAzureUserInput` related types that need to be shared between the agent
extension and the service extensions.
- Create extensions/unions of `IAzureUserInput` related types to include
various types of "agent metadata". Such metadata is designed to help
ensure smooth LLM system prompt creation and execution.
- Clean up the progress updates sent by the slash command handler that
`slashCommandFromWizardBasedExtensionCommand` creates.
- Move the `IAzureUserInput` that
`slashCommandFromWizardBasedExtensionCommand` creates to be its own
class.